+++ SourceryFramework/Sources/Generating/Generator.swift
@@ -7,8 +7,8 @@ import Foundation
 import SourceryRuntime
 
 public enum Generator {
-    public static func generate(_ types: Types, template: Template, arguments: [String: NSObject] = [:]) throws -> String {
+    public static func generate(_ types: Types, functions: [SourceryMethod], template: Template, arguments: [String: NSObject] = [:]) throws -> String {
         Log.verbose("Rendering template \(template.sourcePath)")
-        return try template.render(TemplateContext(types: types, arguments: arguments))
+        return try template.render(TemplateContext(types: types, functions: functions, arguments: arguments))
     }
 }