+++ SourcerySwift/Sources/SourceryRuntime.content.generated.swift
@@ -2055,6 +2055,7 @@ import Foundation
             }
         }
     }
+    public var functions = [SourceryMethod]()
     public var typealiases = [Typealias]()
     public var inlineRanges = [String: NSRange]()
     public var inlineIndentations = [String: String]()
@@ -2062,10 +2063,11 @@ import Foundation
     public var modifiedDate: Date
     public var sourceryVersion: String
 
-    public init(path: String?, module: String?, types: [Type], typealiases: [Typealias] = [], inlineRanges: [String: NSRange] = [:], inlineIndentations: [String: String] = [:], modifiedDate: Date = Date(), sourceryVersion: String = "") {
+    public init(path: String?, module: String?, types: [Type], functions: [SourceryMethod], typealiases: [Typealias] = [], inlineRanges: [String: NSRange] = [:], inlineIndentations: [String: String] = [:], modifiedDate: Date = Date(), sourceryVersion: String = "") {
         self.path = path
         self.module = module
         self.types = types
+        self.functions = functions
         self.typealiases = typealiases
         self.inlineRanges = inlineRanges
         self.inlineIndentations = inlineIndentations
@@ -2081,6 +2083,7 @@ import Foundation
             self.path = aDecoder.decode(forKey: "path")
             self.module = aDecoder.decode(forKey: "module")
             guard let types: [Type] = aDecoder.decode(forKey: "types") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["types"])); fatalError() }; self.types = types
+            guard let functions: [SourceryMethod] = aDecoder.decode(forKey: "functions") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["functions"])); fatalError() }; self.functions = functions
             guard let typealiases: [Typealias] = aDecoder.decode(forKey: "typealiases") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["typealiases"])); fatalError() }; self.typealiases = typealiases
             guard let inlineRanges: [String: NSRange] = aDecoder.decode(forKey: "inlineRanges") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["inlineRanges"])); fatalError() }; self.inlineRanges = inlineRanges
             guard let inlineIndentations: [String: String] = aDecoder.decode(forKey: "inlineIndentations") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["inlineIndentations"])); fatalError() }; self.inlineIndentations = inlineIndentations
@@ -2093,6 +2096,7 @@ import Foundation
             aCoder.encode(self.path, forKey: "path")
             aCoder.encode(self.module, forKey: "module")
             aCoder.encode(self.types, forKey: "types")
+            aCoder.encode(self.functions, forKey: "functions")
             aCoder.encode(self.typealiases, forKey: "typealiases")
             aCoder.encode(self.inlineRanges, forKey: "inlineRanges")
             aCoder.encode(self.inlineIndentations, forKey: "inlineIndentations")
@@ -3291,6 +3295,7 @@ import Foundation
 
 /// :nodoc:
 @objcMembers public final class TemplateContext: NSObject, SourceryModel {
+    public let functions: [SourceryMethod]
     public let types: Types
     public let argument: [String: NSObject]
 
@@ -3299,8 +3304,9 @@ import Foundation
         return types.typesByName
     }
 
-    public init(types: Types, arguments: [String: NSObject]) {
+    public init(types: Types, functions: [SourceryMethod], arguments: [String: NSObject]) {
         self.types = types
+        self.functions = functions
         self.argument = arguments
     }
 
@@ -3308,12 +3314,14 @@ import Foundation
         /// :nodoc:
         required public init?(coder aDecoder: NSCoder) {
             guard let types: Types = aDecoder.decode(forKey: "types") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["types"])); fatalError() }; self.types = types
+            guard let functions: [SourceryMethod] = aDecoder.decode(forKey: "functions") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["functions"])); fatalError() }; self.functions = functions
             guard let argument: [String: NSObject] = aDecoder.decode(forKey: "argument") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["argument"])); fatalError() }; self.argument = argument
         }
 
         /// :nodoc:
         public func encode(with aCoder: NSCoder) {
             aCoder.encode(self.types, forKey: "types")
+            aCoder.encode(self.functions, forKey: "functions")
             aCoder.encode(self.argument, forKey: "argument")
         }
 // sourcery:end
@@ -3321,6 +3329,7 @@ import Foundation
     public var stencilContext: [String: Any] {
         return [
             "types": types,
+            "functions": functions,
             "type": types.typesByName,
             "argument": argument
         ]
@@ -3340,6 +3349,7 @@ import Foundation
                 "inheriting": types.inheriting,
                 "implementing": types.implementing
             ],
+            "functions": functions,
             "type": types.typesByName,
             "argument": argument
         ]