+++ SourceryTests/Generating/JavaScriptTemplateSpecs.swift
@@ -47,7 +47,7 @@ class JavaScriptTemplateTests: QuickSpec {
 
             it("rethrows template parsing errors") {
                 expect {
-                    try Generator.generate(Types(types: []), template: JavaScriptTemplate(templateString: "<% invalid %>"))
+                    try Generator.generate(Types(types: []), functions: [], template: JavaScriptTemplate(templateString: "<% invalid %>"))
                     }
                     .to(throwError(closure: { (error) in
                         expect("\(error)").to(equal(": ReferenceError: ejs:1\n >> 1| <% invalid %>\n\nCan\'t find variable: invalid"))
@@ -56,7 +56,7 @@ class JavaScriptTemplateTests: QuickSpec {
 
             it("rethrows template runtime errors") {
                 expect {
-                    try Generator.generate(Types(types: []), template: JavaScriptTemplate(templateString: "<%_ for (type of types.implementing.Some) { -%><% } %>"))
+                    try Generator.generate(Types(types: []), functions: [], template: JavaScriptTemplate(templateString: "<%_ for (type of types.implementing.Some) { -%><% } %>"))
                     }
                     .to(throwError(closure: { (error) in
                         expect("\(error)").to(equal(": Unknown type Some, should be used with `based`"))
@@ -65,11 +65,21 @@ class JavaScriptTemplateTests: QuickSpec {
 
             it("throws unknown property exception") {
                 expect {
-                    try Generator.generate(Types(types: []), template: JavaScriptTemplate(templateString: "<%_ for (type of types.implements.Some) { -%><% } %>"))
+                    try Generator.generate(Types(types: []), functions: [], template: JavaScriptTemplate(templateString: "<%_ for (type of types.implements.Some) { -%><% } %>"))
                     }
                     .to(throwError(closure: { (error) in
                         expect("\(error)").to(equal(": TypeError: ejs:1\n >> 1| <%_ for (type of types.implements.Some) { -%><% } %>\n\nUnknown property `implements`"))
                     }))
+            }
+
+            it("handles free functions") {
+                let templatePath = Stubs.jsTemplates + Path("Function.ejs")
+                let expectedResult = try? (Stubs.resultDirectory + Path("Function.swift")).read(.utf8)
+
+                expect { try Sourcery(cacheDisabled: true).processFiles(.sources(Paths(include: [Stubs.sourceDirectory])), usingTemplates: Paths(include: [templatePath]), output: output) }.toNot(throwError())
+
+                let result = (try? (outputDir + Sourcery().generatedPath(for: templatePath)).read(.utf8))
+                expect(result).to(equal(expectedResult))
             }
         }
     }