+++ SourceryTests/Parsing/FileParser + MethodsSpec.swift
@@ -17,7 +17,12 @@ class FileParserMethodsSpec: QuickSpec {
             describe("parseMethod") {
                 func parse(_ code: String) -> [Type] {
                     guard let parserResult = try? FileParser(contents: code).parse() else { fail(); return [] }
-                    return Composer.uniqueTypes(parserResult)
+                    return Composer.uniqueTypesAndFunctions(parserResult).types
+                }
+
+                func parseFunctions(_ code: String) -> [SourceryMethod] {
+                    guard let parserResult = try? FileParser(contents: code).parse() else { fail(); return [] }
+                    return Composer.uniqueTypesAndFunctions(parserResult).functions
                 }
 
                 it("extracts methods properly") {
@@ -109,6 +114,12 @@ class FileParserMethodsSpec: QuickSpec {
                             Method(name: "foo()", selectorName: "foo", isStatic: true, definedInTypeName: TypeName("Foo"))
                             ])
                         ]))
+                }
+
+                it("extracts free functions properly") {
+                    expect(parseFunctions("func foo() {}")).to(equal([
+                        Method(name: "foo()", selectorName: "foo", isStatic: false, definedInTypeName: nil)
+                    ]))
                 }
 
                 context("given method with parameters") {