+++ SourceryTests/Stub/Result/Function.swift
@@ -1,0 +1,6 @@
+// Generated using Sourcery Major.Minor.Patch — https://github.com/krzysztofzablocki/Sourcery
+// DO NOT EDIT
+
+func wrappedPerformFoo(value: Foo) {
+    performFoo(value: value)
+}